Little refactoring the statistics plugin
