+++ Configuration/Flexforms/Collection.xml
@@ -22,13 +22,15 @@
                 <el>
                     <settings.collections>
                         <TCEforms>
-                            <displayCond>FIELD:settings.pages:REQ:true</displayCond>
                             <exclude>1</exclude>
                             <label>LLL:EXT:dlf/Resources/Private/Language/locallang_be.xlf:plugins.collection.flexform.collections</label>
                             <config>
                                 <type>select</type>
-                                <items type="array"/>
-                                <itemsProcFunc>Kitodo\Dlf\Hooks\ItemsProcFunc->collectionList</itemsProcFunc>
+                                <renderType>selectMultipleSideBySide</renderType>
+                                <foreign_table>tx_dlf_collections</foreign_table>
+                                <foreign_table_where>AND tx_dlf_collections.hidden = 0 AND tx_dlf_collections.sys_language_uid IN (-1,0)
+                                    ORDER BY tx_dlf_collections.label ASC
+                                </foreign_table_where>
                                 <size>5</size>
                                 <autoSizeMax>15</autoSizeMax>
                                 <maxitems>1024</maxitems>
@@ -39,13 +41,15 @@
                     </settings.collections>
                     <settings.solrcore>
                         <TCEforms>
-                            <displayCond>FIELD:settings.pages:REQ:true</displayCond>
                             <exclude>1</exclude>
                             <label>LLL:EXT:dlf/Resources/Private/Language/locallang_be.xlf:flexform.solrcore</label>
                             <config>
                                 <type>select</type>
-                                <items type="array"></items>
-                                <itemsProcFunc>Kitodo\Dlf\Hooks\ItemsProcFunc->solrList</itemsProcFunc>
+                                <renderType>selectSingle</renderType>
+                                <foreign_table>tx_dlf_solrcores</foreign_table>
+                                <foreign_table_where>ORDER BY
+                                    tx_dlf_solrcores.label ASC
+                                </foreign_table_where>
                                 <size>1</size>
                                 <maxitems>1</maxitems>
                                 <minitems>1</minitems>
