+++ Configuration/Flexforms/Navigation.xml
@@ -20,26 +20,6 @@
                 </TCEforms>
                 <type>array</type>
                 <el>
-                    <settings.pages>
-                        <TCEforms>
-                            <exclude>1</exclude>
-                            <label>LLL:EXT:lang/Resources/Private/Language/locallang_general.xlf:LGL.startingpoint</label>
-                            <config>
-                                <type>group</type>
-                                <internal_type>db</internal_type>
-                                <allowed>pages</allowed>
-                                <size>1</size>
-                                <maxitems>1</maxitems>
-                                <minitems>1</minitems>
-                                <show_thumbs>1</show_thumbs>
-                                <wizards>
-                                    <suggest>
-                                        <type>suggest</type>
-                                    </suggest>
-                                </wizards>
-                            </config>
-                        </TCEforms>
-                    </settings.pages>
                     <settings.features>
                         <TCEforms>
                             <exclude>1</exclude>
