+++ Configuration/Flexforms/PageGrid.xml
@@ -20,26 +20,6 @@
                 </TCEforms>
                 <type>array</type>
                 <el>
-                    <settings.pages>
-                        <TCEforms>
-                            <exclude>1</exclude>
-                            <label>LLL:EXT:lang/Resources/Private/Language/locallang_general.xlf:LGL.startingpoint</label>
-                            <config>
-                                <type>group</type>
-                                <internal_type>db</internal_type>
-                                <allowed>pages</allowed>
-                                <size>1</size>
-                                <maxitems>1</maxitems>
-                                <minitems>1</minitems>
-                                <show_thumbs>1</show_thumbs>
-                                <wizards>
-                                    <suggest>
-                                        <type>suggest</type>
-                                    </suggest>
-                                </wizards>
-                            </config>
-                        </TCEforms>
-                    </settings.pages>
                     <settings.paginate.itemsPerPage>
                         <TCEforms>
                             <exclude>1</exclude>
@@ -85,21 +65,6 @@
                             </config>
                         </TCEforms>
                     </settings.targetPid>
-                    <settings.templateFile>
-                        <TCEforms>
-                            <exclude>1</exclude>
-                            <label>LLL:EXT:dlf/Resources/Private/Language/locallang_be.xlf:flexform.templateFile</label>
-                            <config>
-                                <type>group</type>
-                                <internal_type>file_reference</internal_type>
-                                <allowed>tmpl,tpl,html,htm,txt</allowed>
-                                <size>1</size>
-                                <maxitems>1</maxitems>
-                                <minitems>0</minitems>
-                                <disable_controls>upload</disable_controls>
-                            </config>
-                        </TCEforms>
-                    </settings.templateFile>
                 </el>
             </ROOT>
         </sDEF>
